+++ examples/with-gatsby/README.md
@@ -18,9 +18,9 @@ This Turborepo includes the following packages and apps:
 
 - `docs`: a [Next.js](https://nextjs.org/) app
 - `web`: a [Gatsby.js](https://www.gatsbyjs.com/docs) app
-- `ui`: a stub React component library shared by both `web` and `docs` applications
-- `eslint-config-custom`: `eslint` configurations (includes `eslint-config-next` and `eslint-config-prettier`)
-- `tsconfig`: `tsconfig.json`s used throughout the monorepo
+- `@repo/ui`: a stub React component library shared by both `web` and `docs` applications
+- `@repo/eslint-config`: `eslint` configurations (includes `eslint-config-next` and `eslint-config-prettier`)
+- `@repo/typescript-config`: `tsconfig.json`s used throughout the monorepo
 
 Each package/app is 100% [TypeScript](https://www.typescriptlang.org/).
