+++ examples/with-gatsby/apps/docs/pages/index.tsx
@@ -1,4 +1,4 @@
-import { Button } from "ui/button";
+import { Button } from "@repo/ui/button";
 
 export default function Docs(): JSX.Element {
   return (
