+++ examples/with-gatsby/apps/docs/tsconfig.json
@@ -1,5 +1,5 @@
 {
-  "extends": "tsconfig/nextjs.json",
-  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx"],
+  "extends": "@repo/typescript-config/nextjs.json",
+  "include": ["next-env.d.ts", "next.config.js", "**/*.ts", "**/*.tsx"],
   "exclude": ["node_modules"]
 }
