+++ examples/with-gatsby/apps/web/package.json
@@ -7,22 +7,23 @@
     "build": "gatsby build",
     "start": "gatsby develop",
     "typecheck": "tsc --noEmit",
-    "lint": "eslint src/"
+    "lint": "eslint src/ --max-warnings 0"
   },
   "dependencies": {
     "gatsby": "^5.12.11",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "ui": "workspace:*"
+    "@repo/ui": "workspace:*"
   },
   "devDependencies": {
+    "@repo/eslint-config": "workspace:*",
+    "@repo/typescript-config": "workspace:*",
     "@types/node": "^20.9.4",
     "@types/react": "^18.2.38",
     "@types/react-dom": "^18.2.17",
-    "eslint-config-custom": "workspace:*",
+    "eslint": "^8.54.0",
     "gatsby-plugin-compile-es6-packages": "^2.1.1",
     "gatsby-plugin-pnpm": "^1.2.10",
-    "tsconfig": "workspace:*",
     "typescript": "^5.3.2"
   }
 }
