+++ examples/with-gatsby/apps/web/src/pages/index.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 import type { HeadProps, PageProps } from "gatsby";
-import { Button } from "ui/button";
+import { Button } from "@repo/ui/button";
 
 function IndexPage(_: PageProps): JSX.Element {
   return (
