+++ examples/with-gatsby/apps/web/tsconfig.json
@@ -1,5 +1,5 @@
 {
-  "extends": "tsconfig/gatsby.json",
+  "extends": "@repo/typescript-config/gatsby.json",
   "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx"],
   "exclude": ["node_modules"]
 }
