+++ examples/with-gatsby/package.json
@@ -15,7 +15,6 @@
   "devDependencies": {
     "eslint": "^8.54.0",
     "prettier": "^3.1.0",
-    "tsconfig": "workspace:*",
     "turbo": "^1.10.16"
   },
   "engines": {
