--- examples/with-gatsby/packages/eslint-config-custom/package.json
@@ -1,10 +1,0 @@
-{
-  "name": "eslint-config-custom",
-  "license": "MIT",
-  "version": "0.0.0",
-  "private": true,
-  "devDependencies": {
-    "@vercel/style-guide": "^5.0.0",
-    "eslint-config-turbo": "^1.10.12"
-  }
-}
