+++ examples/with-gatsby/packages/eslint-config/gatsby.js
