+++ examples/with-gatsby/packages/typescript-config/base.json
