+++ examples/with-gatsby/packages/typescript-config/gatsby.json
