+++ examples/with-gatsby/packages/typescript-config/nextjs.json
