+++ examples/with-gatsby/packages/typescript-config/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "tsconfig",
+  "name": "@repo/typescript-config",
   "version": "0.0.0",
   "private": true,
   "files": [
