+++ examples/with-gatsby/packages/typescript-config/react-library.json
