+++ examples/with-gatsby/packages/ui/.eslintrc.js
@@ -1,3 +1,8 @@
+/** @type {import("eslint").Linter.Config} */
 module.exports = {
-  extends: ["custom/react-internal"],
+  extends: ["@repo/eslint-config/react-internal.js"],
+  parser: "@typescript-eslint/parser",
+  parserOptions: {
+    project: true,
+  },
 };
