+++ examples/with-gatsby/packages/ui/package.json
@@ -1,18 +1,19 @@
 {
-  "name": "ui",
+  "name": "@repo/ui",
   "version": "0.0.0",
   "exports": {
     "./button": "./src/button.tsx"
   },
   "scripts": {
-    "lint": "eslint ."
+    "lint": "eslint . --max-warnings 0"
   },
   "devDependencies": {
     "@types/react": "^18.2.38",
     "@types/react-dom": "^18.2.17",
-    "eslint-config-custom": "workspace:*",
+    "@repo/eslint-config": "workspace:*",
+    "@repo/typescript-config": "workspace:*",
+    "eslint": "^8.54.0",
     "react": "^18.2.0",
-    "tsconfig": "workspace:*",
     "typescript": "^5.3.2"
   }
 }
