+++ examples/with-gatsby/packages/ui/src/button.tsx
@@ -1,5 +1,3 @@
-import * as React from "react";
-
 export function Button(): JSX.Element {
   return <button type="button">Boop</button>;
 }
